// apps/gestion-plasticos/helpers/interfaces/create-boleta.dto.ts
export class CreateBoletaDto {
  nombreCliente: string;
  rutCliente: string;
  dirCliente: string;
  bultos: string;
  fecha: Date;
  iva = false;
  detalle: {
    bolsa: string;
    cantidad: string;
    precio: string;
    tipo: 'Normal' | 'Especial';
  }[];
  precio: {
    precioNormal: string;
    precioEspecial: string;
  };
}